fix QolType
